verplaatsing naar bovenbouw
